requirements.txt
Pillow~=8.4.0
setuptools~=57.0.0